@@ -10,11 +10,11 @@
         }
     ],
     "require": {
-        "php": ">=7.1.3"
+        "php": "^8.1"
     },
     "require-dev": {
-        "bolt/core": "^4.0.0",
-        "symplify/easy-coding-standard": "^6.0"
+        "bolt/core": "dev-5.2.0-rc.2",
+        "symplify/easy-coding-standard": "^12.2"
     },
     "autoload": {
         "psr-4": {
@@ -28,5 +28,45 @@
         "screenshots": [
             "screenshots/widget.png"
         ]
+    },
+    "repositories": {
+        "antiftw-bolt-core": {
+            "type": "vcs",
+            "url": "https://github.com/antiftw/bolt-core"
+        },
+        "jfsenechal_doctrine_behaviors": {
+            "type": "vcs",
+            "url": "git@github.com:jfsenechal/DoctrineBehaviors.git"
+        },
+        "antiftw-yaml-migrations": {
+            "type": "vcs",
+            "url": "https://github.com/antiftw/bdo-yaml-migrations"
+        },
+        "antiftw-php-translation-symfony-bundle": {
+            "type": "vcs",
+            "url": "https://github.com/antiftw/php-translation-symfony-bundle"
+        },
+        "antiftw-php-translation-extractor": {
+            "type": "vcs",
+            "url": "https://github.com/antiftw/php-translation-extractor"
+        },
+        "antiftw-php-translation-loco-adapter": {
+            "type": "vcs",
+            "url": "https://github.com/antiftw/php-translation-loco-adapter"
+        },
+        "antiftw-bolt-reference-extension": {
+            "type": "vcs",
+            "url": "https://github.com/antiftw/bolt-reference-extension"
+        },
+        "antiftw-bdo-configuration-notices": {
+            "type": "vcs",
+            "url": "https://github.com/antiftw/bdo-configuration-notices"
+        }
+    },
+    "config": {
+        "allow-plugins": {
+            "symfony/flex": true,
+            "drupol/composer-packages": true
+        }
     }
 }
